cms优化栏目管理

+++ jbone-cms/jbone-cms-common/src/main/java/cn/jbone/cms/common/dataobject/CategoryTocDO.java
@@ -15,7 +15,7 @@ public class CategoryTocDO {
     /**
      * 父级栏目
      */
-    private int pid;
+    private long pid;
 
     /**
      * 栏目标题
